+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2025-02-10]
+## [0.0.1 - 2025-02-11]
 ### Added
 - First version of the project
 - Spring Application
@@ -215,3 +215,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Log DataSHIELD Token Manager Job and Service
 - Log Coder Service and Job
 - Log App Register Service
+- Removal conditions for ProjectBridgeheadDataShield